RAG_GEMINI/requirements.txt
streamlit==1.28.0
numpy==1.24.3
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
google-genai==0.2.0